Move estimateGas tests to seperate file
